feat(standalone-mode): Enable using composite controller as a library
Currently it is only possible to run Metacontroller as a cluster-wide
singletone. This mode requires deploying two additional CRDs,
CompositeController and DecoratorController.

This commit makes the changes needed to use composite controller
as a library from third-party applications. It introduces the standalone
mode and adds an example of how it can be used.

Signed-off-by: fpetkovski <filip.petkovsky@gmail.com>